pyproject.toml
[project]
name = "mci-cot"
version = "0.1.0"
description = "MCI-CoT experiments with Hydra"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "peft>=0.10.0",
  "hydra-core>=1.3.2",
  "wandb>=0.16.0",
  "optuna>=3.6.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "pandas>=2.1.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "tqdm>=4.66.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
